+++ libs/community/langchain_community/tools/azure_ai_services/document_intelligence.py
